Feature/cucumber parallel

+++ lib/report_portal/cucumber/parallel_report.rb
@@ -1,78 +1,60 @@
-# Copyright 2015 EPAM Systems
-# 
-# 
-# This file is part of Report Portal.
-# 
-# Report Portal is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-# 
-# ReportPortal is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser General Public License for more details.
-# 
-# You should have received a copy of the GNU Lesser General Public License
-# along with Report Portal.  If not, see <http://www.gnu.org/licenses/>.
-
 require 'parallel_tests'
 require 'sys/proctable'
 require 'fileutils'
-
 require_relative 'report'
 
 module ReportPortal
   module Cucumber
     class ParallelReport < Report
 
-      def parallel?
-        true
-      end
 
-      def initialize(desired_time)
-        @root_node = Tree::TreeNode.new('')
+      def initialize
         ReportPortal.last_used_time = 0
+        @root_node = Tree::TreeNode.new('')
+        @parent_item_node = @root_node
+
         set_parallel_tests_vars
+
         if ParallelTests.first_process?
-          start_launch(desired_time)
+          start_launch(ReportPortal.now)
         else
           start_time = monotonic_time
           loop do
             break if File.exist?(lock_file)
-            if monotonic_time - start_time > wait_time_for_launch_start
-              raise "File with launch ID wasn't created after waiting #{wait_time_for_launch_start} seconds"
+            if monotonic_time - start_time > wait_time_for_launch_create
+              raise "File with launch ID wasn't created after waiting #{wait_time_for_launch_create} seconds"
             end
+
             sleep 0.5
           end
-          sleep_time = 5
-          sleep(sleep_time) # stagger start times for reporting to Report Portal to avoid collision
           ReportPortal.launch_id = read_lock_file(lock_file)
           add_process_description
         end
       end
 
       def add_process_description
-        description = ReportPortal.get_launch['description'].split(' ')
-        description.push(self.description().split(' ')).flatten!
-        ReportPortal.update_launch({description: description.uniq.join(' ')})
+        description = ReportPortal.remote_launch['description'].split(' ')
+        description.push(self.description.split(' ')).flatten!
+        ReportPortal.update_launch(description: description.uniq.join(' '))
       end
 
-      def done(desired_time = ReportPortal.now)
-        end_feature(desired_time) if @feature_node
+      def test_run_finished(_event, desired_time )
+        end_feature(desired_time) unless @parent_item_node.is_root?
 
-        if ParallelTests.first_process?
-          ParallelTests.wait_for_other_processes_to_finish
+        if parallel
+          if ParallelTests.first_process?
+            ParallelTests.wait_for_other_processes_to_finish
+            close_all_children_of(@root_node) # Folder items are closed here as they can't be closed after finishing a feature
 
-          File.delete(lock_file)
+            File.delete(lock_file)
 
-          # TODO: if parallel test does not need to finish launch, there should be env variable to support this.
-          #  as report launch is is created by first process it should also be finished by same process
-          unless ReportPortal::Settings.instance.launch_id || ReportPortal::Settings.instance.file_with_launch_id
-            ReportPortal.close_child_items(nil)
-            time_to_send = time_to_send(desired_time)
-            ReportPortal.finish_launch(time_to_send)
+            if started_launch || !attach_to_launch?
+              time_to_send = time_to_send(desired_time)
+              ReportPortal.finish_launch(time_to_send)
+            end
           end
+        else
+          super(_event,desired_time)
         end
       end
 
@@ -88,27 +70,30 @@ module ReportPortal
         runner_process ||= get_parallel_test_process(process_list)
         runner_process ||= get_cucumber_test_process(process_list)
         raise 'Could not find parallel_cucumber/parallel_test in ancestors of current process' if runner_process.nil?
+
         @pid_of_parallel_tests = runner_process.pid
-        @cmd_args_of_parallel_tests = runner_process.cmdline.split(' ',2).pop
+        @cmd_args_of_parallel_tests = runner_process.cmdline.split(' ', 2).pop
       end
 
       def get_parallel_test_process(process_list)
         process_list.each do |process|
-          return process if process.cmdline.match(/bin(?:\/|\\)parallel_(?:cucumber|test)(.+)/)
+          if process.cmdline.match(%r{bin(?:\/|\\)parallel_(?:cucumber|test)(.+)})
+            @parallel = true
+            return process
+          end
         end
         nil
       end
 
       def get_cucumber_test_process(process_list)
         process_list.each do |process|
-          return process if process.cmdline.match(/bin(?:\/|\\)(?:cucumber)(.+)/)
+          return process if process.cmdline.match(%r{bin(?:\/|\\)(?:cucumber)(.+)})
         end
         nil
       end
 
-      #time required for first tread to created remote project in RP and save id to file
-      def wait_time_for_launch_start
-        ENV['rp_parallel_launch_wait_time'] ? ENV['rp_parallel_launch_wait_time'] : 60
+      def wait_time_for_launch_create
+        ENV['rp_parallel_launch_wait_time'] || 60
       end
 
       def monotonic_time
